refactor(work-control): split runtime responsibilities
