+++ .buildchain/kfd/kfd-2/release-claims.json
@@ -11,7 +11,7 @@
     "version": "4.0.0-alpha.4",
     "channel": "repository-default",
     "tag": "v4.0.0-alpha.4",
-    "sourceSha": "2d5aa4338d67f7696c790e920d86b0fc4cbd111d"
+    "sourceSha": "e4f22f747576f18c7f6adea908c64ca532f8eddc"
   },
   "claims": [
     {
