+++ framework/site/src/kfx-site-impact-proofs/b00970ee90e6ed4eff9735f09c349c0d7d1af0608749609eb2478bd0ee4f9036.json
@@ -1,0 +1,154 @@
+{
+  "contract": "kungfu.kfx-site-impact-proof/v1",
+  "baseRevision": "e4f22f747576f18c7f6adea908c64ca532f8eddc",
+  "changeRoot": "sha256:14a7db711d4143194aa707a7210ee89b36f5108122e058d56a395672ec3cd85a",
+  "changes": [
+    {
+      "baseContentRoot": "sha256:aee825a02297456cbcf454edbc77e069050ed6ed4eccd4bcf2c3d1761b48a409",
+      "contentRoot": "sha256:c52cc5b5ca1bc16bdd2773a974089e354392608243dc9f56956f9d469ee47f3c",
+      "path": ".buildchain/kfd/kfd-1/contract-world.witness.json",
+      "status": "modified"
+    },
+    {
+      "baseContentRoot": "sha256:6fae25376347e519e97d4c40c67f9d294d237d06b4297c164f2077638e49281d",
+      "contentRoot": "sha256:f1c8713be33ddff0b544c900122cfaca65cb29c65decf21289cd826d396548e1",
+      "path": ".buildchain/kfd/kfd-1/release-gate.json",
+      "status": "modified"
+    },
+    {
+      "baseContentRoot": "sha256:2cee4c24d1dd1ab9550110f2a5f211a74c8e3d159193590a35889c17a36578f1",
+      "contentRoot": "sha256:985d0e522e6b3a2aeec5a77329633745538a4a5fd4476a6c546ef045bd6c6814",
+      "path": ".buildchain/kfd/kfd-2/release-claims.json",
+      "status": "modified"
+    },
+    {
+      "baseContentRoot": "sha256:7f913c8a365569f612ff5f4056bdd12a9c2960a9d3f32e0d9fb8372fe72d379a",
+      "contentRoot": "sha256:3c542f198e0aab7642fc1d30179898d56e9d11b1055f6d09e2deeb7dd91d9146",
+      "path": ".buildchain/kfd/kfd-3/collaboration-interface.prebuild.json",
+      "status": "modified"
+    },
+    {
+      "baseContentRoot": "sha256:60baf341ea6916c8ef662ce622701059b6bf013fc08918a56bc33d195cb06ae3",
+      "contentRoot": "sha256:4358e26a6e311350fe92c96335b8b4344dc31a3ed15048d50b7b4bef3fada369",
+      "path": ".buildchain/kfd/support-matrix.json",
+      "status": "modified"
+    },
+    {
+      "baseContentRoot": "sha256:0daf4dfd723063ea730af0351c3cebe41b939a6b7ab3e185e4e6d232f3043844",
+      "contentRoot": "sha256:36e272a5a99b0644edeea19ecf7d1968802fb44ae28e874f43edbdfcb8fd7ced",
+      "path": ".xinfa/portable-atlas-classification.json.gz",
+      "status": "modified"
+    },
+    {
+      "baseContentRoot": "sha256:c89a436b5ab99b89efd6057dd22eeac46c1cd0f8fa2fc5ae9b9c7fa18ca20a0d",
+      "contentRoot": "sha256:9265f14a50ce4f791d447f84bd2b0e0d89c31bd4b63a04c7cef79bf48dde4851",
+      "path": ".xinfa/product-atlas-bundle.json",
+      "status": "modified"
+    },
+    {
+      "baseContentRoot": "sha256:6b095ec2dc8b5612ab590e2e173e68eba7897e484d7ed9ab7054b9989ae898ce",
+      "contentRoot": "sha256:5edc4ae64061536ec525d5dc569493ca967f8044a7b2e8211df48161189f9513",
+      "path": "config/kungfu-agent-first-canonical-policy.json",
+      "status": "modified"
+    },
+    {
+      "baseContentRoot": "sha256:bf21a69efe053bb0a5ec2e2458f22ef9f14fc44ba94dc617cd8c22fe5d4c02fe",
+      "contentRoot": "sha256:a7a10d4b6bf9cef652d5e3bedd094ab68a7572cc6e3e66d6a53812c4c7894a3b",
+      "path": "config/primitive/kungfu-primitive-catalog.contract.json",
+      "status": "modified"
+    },
+    {
+      "baseContentRoot": "sha256:aee825a02297456cbcf454edbc77e069050ed6ed4eccd4bcf2c3d1761b48a409",
+      "contentRoot": "sha256:c52cc5b5ca1bc16bdd2773a974089e354392608243dc9f56956f9d469ee47f3c",
+      "path": "developer/sdk/kfd/kfd-1/contract-world.witness.json",
+      "status": "modified"
+    },
+    {
+      "baseContentRoot": "sha256:6fae25376347e519e97d4c40c67f9d294d237d06b4297c164f2077638e49281d",
+      "contentRoot": "sha256:f1c8713be33ddff0b544c900122cfaca65cb29c65decf21289cd826d396548e1",
+      "path": "developer/sdk/kfd/kfd-1/release-gate.json",
+      "status": "modified"
+    },
+    {
+      "baseContentRoot": "sha256:2cee4c24d1dd1ab9550110f2a5f211a74c8e3d159193590a35889c17a36578f1",
+      "contentRoot": "sha256:985d0e522e6b3a2aeec5a77329633745538a4a5fd4476a6c546ef045bd6c6814",
+      "path": "developer/sdk/kfd/kfd-2/release-claims.json",
+      "status": "modified"
+    },
+    {
+      "baseContentRoot": "sha256:60baf341ea6916c8ef662ce622701059b6bf013fc08918a56bc33d195cb06ae3",
+      "contentRoot": "sha256:4358e26a6e311350fe92c96335b8b4344dc31a3ed15048d50b7b4bef3fada369",
+      "path": "developer/sdk/kfd/support-matrix.json",
+      "status": "modified"
+    },
+    {
+      "contentRoot": "sha256:968890ad62206df355ca2a541213ae1307311e8b2906e0150ce75496bb94714b",
+      "path": "extensions/work-control/work-control-actions/domain/_work_control_assignment.py",
+      "status": "added"
+    },
+    {
+      "contentRoot": "sha256:3821a5815bd1d149141e1ed65774418affe40398505233c540d38c3a49d27738",
+      "path": "extensions/work-control/work-control-actions/domain/_work_control_lifecycle.py",
+      "status": "added"
+    },
+    {
+      "contentRoot": "sha256:8776fe6ba458b9c9aa6888a240f20ababc9056a03185624d96f77c6e1282902a",
+      "path": "extensions/work-control/work-control-actions/domain/_work_control_state.py",
+      "status": "added"
+    },
+    {
+      "baseContentRoot": "sha256:a2e64cf0376b5bb639f35adbda603eef0eb75ad033503c84d47d2efdb2213692",
+      "contentRoot": "sha256:3360f81fbd30aa42ff4889d53e58c4608f4c2ec8dbcd8207d36fb920c7f04746",
+      "path": "extensions/work-control/work-control-actions/domain/work_control_runtime.py",
+      "status": "modified"
+    },
+    {
+      "baseContentRoot": "sha256:e50d95043f99770df399e87adc2d5ac846f127c88a45f9f947978c9bb1ccb9fc",
+      "contentRoot": "sha256:33b7c7c8a9323b98838382034bf6288974fd18d6fbdf6359fdb655d98534483c",
+      "path": "framework/assignment-runtime/assignment-runtime.test.mjs",
+      "status": "modified"
+    },
+    {
+      "baseContentRoot": "sha256:ba3c648af84ab2dc81b4b43268dfc8d954f9895b6d05436cde2cbd9279e15444",
+      "contentRoot": "sha256:e2736dd3efd185e9ed9c58937f98edc4f37a170b87b64d5453272dd48e13f7dd",
+      "path": "framework/assignment-runtime/work-authority-topology-v1.json",
+      "status": "modified"
+    },
+    {
+      "baseContentRoot": "sha256:6b095ec2dc8b5612ab590e2e173e68eba7897e484d7ed9ab7054b9989ae898ce",
+      "contentRoot": "sha256:5edc4ae64061536ec525d5dc569493ca967f8044a7b2e8211df48161189f9513",
+      "path": "framework/contract/kungfu-agent-first-canonical-policy.json",
+      "status": "modified"
+    },
+    {
+      "baseContentRoot": "sha256:f875813988d12efe8847fa7b0cf0605d2fac4edf33a87eb35552c491b9b28f29",
+      "contentRoot": "sha256:199fcfeae220bf9a93bf7e4ecbada2c43e3806db9cea8a667d1e01f42aca756b",
+      "path": "framework/core/src/libkungfu/include/kungfu/sdk/generated/primitive_catalog_v2.hpp",
+      "status": "modified"
+    },
+    {
+      "baseContentRoot": "sha256:bf21a69efe053bb0a5ec2e2458f22ef9f14fc44ba94dc617cd8c22fe5d4c02fe",
+      "contentRoot": "sha256:a7a10d4b6bf9cef652d5e3bedd094ab68a7572cc6e3e66d6a53812c4c7894a3b",
+      "path": "framework/primitive/kungfu-primitive-catalog.contract.json",
+      "status": "modified"
+    },
+    {
+      "baseContentRoot": "sha256:f3d197e3e53f15b3d3728ef2a9eae424daf9aab94aeab4f24ebbc6e7ae6798e8",
+      "contentRoot": "sha256:26662d79e7e1ff4f62f5ac0361e41c5b20c8ddccb9e7130b288e0692de867b13",
+      "path": "package.json",
+      "status": "modified"
+    }
+  ],
+  "affectedFacets": [
+    "capabilities",
+    "package-facets",
+    "profiles-suites"
+  ],
+  "rationale": "Work Control runtime implementation was partitioned behind the unchanged Python facade; Assignment authority topology and tests were updated, while KFX Site product behavior and public extension surfaces remain unchanged.",
+  "attestation": {
+    "publicBehaviorUnchanged": true,
+    "publicContractUnchanged": true,
+    "readerJourneyUnchanged": true
+  },
+  "proofRoot": "sha256:b00970ee90e6ed4eff9735f09c349c0d7d1af0608749609eb2478bd0ee4f9036"
+}
