+++ package.json
@@ -163,7 +163,7 @@
     "test:health-preflight-performance": "node scripts/require-shifu.mjs test:health-preflight-performance && node framework/core/tests/qualification/health-preflight/run.mjs",
     "test:action-kernel": "node scripts/require-shifu.mjs test:action-kernel && node --test framework/action/action.test.mjs framework/action/action-loop-contract.test.mjs framework/action/action-loop-begin.test.mjs framework/action/action-loop-settle.test.mjs && uv run --project framework/core --frozen pytest -q framework/core/tests/python/test_action_command.py",
     "test:assignment-capture": "node scripts/require-shifu.mjs test:assignment-capture && node --test framework/assignment-capture/assignment-capture.test.mjs",
-    "test:assignment-runtime": "node scripts/require-shifu.mjs test:assignment-runtime && node --test framework/assignment-runtime/assignment-runtime.test.mjs framework/api/tests/assignment-runtime.test.ts framework/gui/src/main/assignment-runtime-host.test.ts framework/gui/src/product-search.test.ts extensions/work-dashboard/tests/work-control-profile.test.ts && PYTHONPATH=framework/core/src/python:${KUNGFU_NATIVE_PATH:-framework/core/build/Release} uv run --project framework/core --frozen ruff format --check framework/core/src/python/kungfu/assignment_runtime framework/core/src/python/kungfu/cli/commands/assignment.py framework/core/src/python/kungfu/coordination/locks.py framework/core/src/python/kungfu/runtime_service.py extensions/work-control/work-control-actions/adapter.py extensions/work-control/work-control-actions/domain/work_control_runtime.py extensions/work-control/work-control-actions/domain/work_semantics.py framework/core/tests/python/test_assignment_runtime.py framework/core/tests/python/test_coordination_locks.py framework/core/tests/python/test_runtime_service.py && PYTHONPATH=framework/core/src/python:${KUNGFU_NATIVE_PATH:-framework/core/build/Release} uv run --project framework/core --frozen pytest -q framework/core/tests/python/test_assignment_runtime.py framework/core/tests/python/test_coordination_locks.py && PYTHONPATH=framework/core/src/python:${KUNGFU_NATIVE_PATH:-framework/core/build/Release} uv run --project framework/core --frozen pytest -q framework/core/tests/python/test_runtime_service.py -k 'process_runtime_host_never_reports_ready_when_native_bind_fails or wait_for_coordinator'",
+    "test:assignment-runtime": "node scripts/require-shifu.mjs test:assignment-runtime && node --test framework/assignment-runtime/assignment-runtime.test.mjs framework/api/tests/assignment-runtime.test.ts framework/gui/src/main/assignment-runtime-host.test.ts framework/gui/src/product-search.test.ts extensions/work-dashboard/tests/work-control-profile.test.ts && PYTHONPATH=framework/core/src/python:${KUNGFU_NATIVE_PATH:-framework/core/build/Release} uv run --project framework/core --frozen ruff format --check framework/core/src/python/kungfu/assignment_runtime framework/core/src/python/kungfu/cli/commands/assignment.py framework/core/src/python/kungfu/coordination/locks.py framework/core/src/python/kungfu/runtime_service.py extensions/work-control/work-control-actions/adapter.py extensions/work-control/work-control-actions/domain framework/core/tests/python/test_assignment_runtime.py framework/core/tests/python/test_coordination_locks.py framework/core/tests/python/test_runtime_service.py && PYTHONPATH=framework/core/src/python:${KUNGFU_NATIVE_PATH:-framework/core/build/Release} uv run --project framework/core --frozen pytest -q framework/core/tests/python/test_assignment_runtime.py framework/core/tests/python/test_coordination_locks.py && PYTHONPATH=framework/core/src/python:${KUNGFU_NATIVE_PATH:-framework/core/build/Release} uv run --project framework/core --frozen pytest -q framework/core/tests/python/test_runtime_service.py -k 'process_runtime_host_never_reports_ready_when_native_bind_fails or wait_for_coordinator'",
     "test:assignment-runtime-gui": "node scripts/require-shifu.mjs test:assignment-runtime-gui && node --test framework/api/tests/assignment-runtime.test.ts framework/gui/src/main/assignment-runtime-host.test.ts framework/gui/src/product-search.test.ts extensions/work-dashboard/tests/work-control-profile.test.ts",
     "test:action-loop-native-authority": "node scripts/require-shifu.mjs test:action-loop-native-authority && node scripts/run-action-loop-native-authority-tests.mjs",
     "test:action-primitive-cli": "node scripts/require-shifu.mjs test:action-primitive-cli && uv run --project framework/core --frozen ruff format --check framework/core/src/python/kungfu/cli/commands/xinfa.py framework/core/src/python/kungfu/cli/commands/primitive_role.py framework/core/src/python/kungfu/cli/commands/pursuit.py framework/core/src/python/kungfu/cli/commands/warrant.py framework/core/src/python/kungfu/cli/commands/episode.py framework/core/tests/python/test_xinfa_command.py framework/core/tests/python/test_action_primitive_role_commands.py && uv run --project framework/core --frozen pytest -q framework/core/tests/python/test_xinfa_command.py framework/core/tests/python/test_action_primitive_role_commands.py framework/core/tests/python/test_action_command.py",
